Module robots::actors

Actors core.

Reexports

pub use self::actor_cell::ActorCell;

pub use self::actor_cell::ActorContext;

pub use self::actor_cell::ControlMessage;

pub use self::actor_cell::InnerMessage;

pub use self::actor_cell::SystemMessage;

pub use self::actor_ref::ActorPath;

pub use self::actor_ref::ActorRef;

pub use self::actor_system::ActorSystem;

pub use self::props::Props;

Modules

actor_cell
Module with the internals of Actors.

actor_ref
Module for ActorRef and CanReceive, the interface given to the user to interract with actors.

actor_system
Module for the ActorSystem.

props
Module for Actor factories (Props).

Traits

Actor
This is the trait to implement to become an Actor.

Any
A type to emulate dynamic typing.

Arguments
Trait to be implemented by args, this is automatically given if a struct is already Clone + Send + Sync + 'static + Any.

Message
Trait to be implemented by messages, this is automatically given if a struct is already Clone + Send + Sync + 'static + Any.