Struct robots::actors::actor_cell::ActorCell

pub struct ActorCell { /* fields omitted */ }

Main interface for interractiong with an Actor for the internals.

Methods

impl ActorCell

fn new(props: Arc<ActorFactory>,
system: ActorSystem,
father: ActorRef,
path: Arc<ActorPath>)
-> ActorCell

Creates a new ActorCell.

fn receive_message(&self, message: InnerMessage, sender: ActorRef)

Puts a message with its sender in the Actor's mailbox and schedules the Actor.

fn receive_system_message(&self, system_message: SystemMessage)

Puts a system message with its sender in the Actor's system mailbox and schedules the Actor.

fn handle_envelope(&self)

Makes the Actor handle an envelope in its mailbox.