Trait robots::actors::actor_cell::ActorContext

pub trait ActorContext<Args: Message, M: Message, A: Actor<M> + 'static> {
    fn actor_ref(&self) -> Arc<ActorRef<Args, M, A>>;
    fn actor_of<ArgsBis: Message, MBis: Message, ABis: Actor<MBis> + 'static>(&self, props: Props<ArgsBis, MBis, ABis>, name: String) -> Arc<ActorRef<ArgsBis, MBis, ABis>>;
    fn tell<MessageTo: Message>(&self, to: Arc<CanReceive>, message: MessageTo);
    fn stop(&self, actor_ref: Arc<CanReceive>);
    fn kill_me(&self);
    fn sender(&self) -> Arc<CanReceive>;
    fn father(&self) -> Arc<CanReceive>;
    fn children(&self) -> Vec<Arc<CanReceive>>;
    fn monitoring(&self) -> Vec<Arc<CanReceive>>;
    fn path(&self) -> Arc<String>;
}

This is the API that Actors are supposed to see of their context while handling a message.

Required Methods

fn actor_ref(&self) -> Arc<ActorRef<Args, M, A>>

Returns an ActorRef of the Actor.

fn actor_of<ArgsBis: Message, MBis: Message, ABis: Actor<MBis> + 'static>(&self, props: Props<ArgsBis, MBis, ABis>, name: String) -> Arc<ActorRef<ArgsBis, MBis, ABis>>

Spawns an actor.

Note that the supervision is not yet implemented so it does the same as creating an actor through the actor system.

fn tell<MessageTo: Message>(&self, to: Arc<CanReceive>, message: MessageTo)

Sends a Message to the targeted CanReceive.

fn stop(&self, actor_ref: Arc<CanReceive>)

Requests the targeted actor to stop.

fn kill_me(&self)

Asks the father to kill the actor.

fn sender(&self) -> Arc<CanReceive>

Returns an Arc to the sender of the message being handled.

fn father(&self) -> Arc<CanReceive>

Father of the actor.

fn children(&self) -> Vec<Arc<CanReceive>>

Children of the actor.

fn monitoring(&self) -> Vec<Arc<CanReceive>>

Lifecycle monitoring, list of monitored actors.

fn path(&self) -> Arc<String>

Lifecycle monitoring, list of monitored actors.