Struct robots::actors::actor_ref::ActorRef

pub struct ActorRef<Args: Message, M: Message, A: Actor<M> + 'static> {
    // some fields omitted
}

This is a reference to an Actor and what is supposed to be manipulated by the user.

The only thing it can do is send and receive messages (according to the actor model in defined by Hewitt).

Methods

impl<Args: Message, M: Message, A: Actor<M> + 'static> ActorRef<Args, M, A>

fn with_cell(cell: ActorCell<Args, M, A>, path: ActorPath) -> ActorRef<Args, M, A>

Creates an ActorRef with the given ActorCell.

fn tell_to<MessageTo: Message>(&self, to: Arc<CanReceive>, message: MessageTo)

Sends a Message to a CanReceive.

fn ask_to<MessageTo: Message, V: Message, E: Send + 'static>(&self, to: Arc<CanReceive>, message: MessageTo) -> Future<V, E>

Sends a Message to a CanReceive.