Trait robots::actors::actor_ref::CanReceive

pub trait CanReceive: Send + Sync {
    fn receive(&self, message: Box<Any>, sender: Arc<CanReceive>);
    fn receive_system_message(&self, system_message: SystemMessage);
    fn handle(&self);
    fn path(&self) -> ActorPath;

    fn equals(&self, other: &CanReceive) -> bool { ... }
}

Trait used to signal that a struct can receive messages. Note that for the moment these are not typed, but it will be easy to add.

Required Methods

fn receive(&self, message: Box<Any>, sender: Arc<CanReceive>)

Puts the message in a mailbox and enqueues the CanReceive.

fn receive_system_message(&self, system_message: SystemMessage)

Puts the system message in a mailbox and enqueues the CanReceive.

fn handle(&self)

Handles the message.

Thus completes a Promise or calls the Actor's receive method.

fn path(&self) -> ActorPath

Path to the actor.

Provided Methods

fn equals(&self, other: &CanReceive) -> bool

Tool for comparing actor refs.