Struct robots::actors::actor_system::ActorSystem

pub struct ActorSystem {
    // some fields omitted
}

ActorSystem, the struct that manages the creation of everything and that everything does what it is supposed to do.

Methods

impl ActorSystem

fn new(name: String) -> ActorSystem

Creates a new ActorSystem.

Note that no threads are started.

fn actor_of<Args: Message, M: Message, A: Actor<M> + 'static>(&self, props: Props<Args, M, A>, name: String) -> Arc<ActorRef<Args, M, A>>

Spawns an Actor of type A, created using the Props given.

fn enqueue_actor<Args: Message, M: Message, A: Actor<M> + 'static>(&self, actor_ref: Arc<ActorRef<Args, M, A>>)

Enqueues the given Actor on the queue of Actors with something to handle.

fn spawn_thread(&self)

Spawns a thread that will consume messages from the ActorRef in actors_queue. This thread can be terminated by calling terminate_thread.

fn terminate_thread(&self)

Kills a consumer thread of the ActorSystem.

fn spawn_threads(&self, n: u32)

Spawns n threads that will consume messages from the ActorRef in actors_queue.

fn terminate_threads(&self, n: u32)

Kills n consumer threads.