Struct robots::actors::props::Props

pub struct Props<Args: Copy + Sync, M: Message, A: Actor<M>> {
    // some fields omitted
}

Factory for A.

It will always create an A with the same function and arguments.

It is also thread safe, and thus we can respawn an Actor across different threads.

Methods

impl<Args: Copy + Sync, M: Message, A: Actor<M>> Props<Args, M, A>

fn new(creator: Arc<Fn(Args) -> A + Sync + Send>, args: Args) -> Props<Args, M, A>

Creates a Props which is a factory for A with the creator function and args args.

fn create(&self) -> A

Creates an Actor instance with the creator function and the args args.

This is meant to allow to respawn an Actor when it fails.