Trait robots::actors::Actor

pub trait Actor<M: Message>: Send + Sync + Sized {
    fn receive<Args: Message>(&self, message: M, context: ActorCell<Args, M, Self>);

    fn receive_termination<Args: Message>(&self, _context: ActorCell<Args, M, Self>) { ... }
    fn pre_start<Args: Message>(&self, _context: ActorCell<Args, M, Self>) { ... }
    fn post_stop(&self) { ... }
    fn pre_restart<Args: Message>(&self, _context: ActorCell<Args, M, Self>) { ... }
    fn post_restart<Args: Message>(&self, context: ActorCell<Args, M, Self>) { ... }
}

This is the trait to implement to become an Actor.

Normaly only the receive method has to be implemented.

Required Methods

fn receive<Args: Message>(&self, message: M, context: ActorCell<Args, M, Self>)

Single method to be implemented for an Actor.

This defines the Actor's behaviour.

Note that Actors have to be both Send AND Sync: - The Send is obvious as they go around threads. - The Sync is needed because of Actor failure handling.
Actors need to be stored in InnerActorCell in a container that offers inner mutability because of the ability to restart them (we replace the old failed actor with a new clean one).
Two container offer that in Sync way, RwLock and Mutex. Mutex would give Sync to the Actor, but if a thread fails while holding a lock, the mutex becomes poisoned and there is no way to unpoison him, so we can't use one. We thus have to use an RwLock, but it does not gives the Sync trait for free, the contained object needs to be Sync itself, thus forcing Actors to be Sync.

extern crate robots; use robots::{Actor, ActorCell, Message};

struct MyActor;

impl Actor for MyActor { fn receive (&self, message: Message, _context: ActorCell) { match message { Message::Dummy => context.tell(context.sender(), Message::Dummy), _ => println!("Hi !"), } } }

Provided Methods

fn receive_termination<Args: Message>(&self, _context: ActorCell<Args, M, Self>)

Method called when a monitored actor is terminated. This is put in a separated method because match in rust must check all variations and we chose not to force the user to make a case for terminations if it doesn not monitor any actor.

fn pre_start<Args: Message>(&self, _context: ActorCell<Args, M, Self>)

Method called before the Actor is started.

fn post_stop(&self)

Method called after the Actor is stopped.

fn pre_restart<Args: Message>(&self, _context: ActorCell<Args, M, Self>)

Method called before the Actor is restarted.

fn post_restart<Args: Message>(&self, context: ActorCell<Args, M, Self>)

Method called after the Actor is restarted.