Trait robots::actors::Arguments

pub trait Arguments: Clone + Send + Sync + 'static { }

Trait to be implemented by args, this is automatically given if a struct is already Clone + Send + Sync + 'static + Any.