Struct robots::actors::actor_cell::ActorCell

pub struct ActorCell<Args: Arguments, A: Actor + 'static> {
    // some fields omitted
}

Main interface for accessing the main Actor information (system, mailbox, sender, props...).

Methods

impl<Args: Arguments, A: Actor + 'static> ActorCell<Args, A>

fn new(actor: A, props: Props<Args, A>, system: ActorSystem, father: Arc<CanReceive>, name: Arc<String>, path: ActorPath) -> ActorCell<Args, A>

Creates a new ActorCell.

fn receive_message(&self, message: InnerMessage, sender: Arc<CanReceive>)

Puts a message with its sender in the Actor's mailbox and schedules the Actor.

fn receive_system_message(&self, system_message: SystemMessage)

Puts a system message with its sender in the Actor's system mailbox and schedules the Actor.

fn handle_envelope(&self)

Makes the Actor handle an envelope in its mailbox.