Trait robots::actors::actor_cell::ActorContext

pub trait ActorContext {
    fn actor_ref(&self) -> ActorRef;
    fn actor_of(&self, props: Arc<ActorFactory>, name: String) -> ActorRef;
    fn tell<MessageTo: Message>(&self, to: ActorRef, message: MessageTo);
    fn stop(&self, actor_ref: ActorRef);
    fn kill_me(&self);
    fn sender(&self) -> ActorRef;
    fn father(&self) -> ActorRef;
    fn children(&self) -> Vec<ActorRef>;
    fn monitoring(&self) -> Vec<ActorRef>;
    fn path(&self) -> Arc<ActorPath>;
    fn identify_actor(&self, _name: String) -> Future<Option<ActorRef>, &'static str>;
}

This is the API that Actors are supposed to see of their context while handling a message.

Required Methods

fn actor_ref(&self) -> ActorRef

Returns an ActorRef to the Actor.

fn actor_of(&self, props: Arc<ActorFactory>, name: String) -> ActorRef

Spawns a child actor.

fn tell<MessageTo: Message>(&self, to: ActorRef, message: MessageTo)

Sends a Message to the targeted ActorRef.

fn stop(&self, actor_ref: ActorRef)

Requests the targeted actor to stop.

fn kill_me(&self)

Asks the father of the actor to terminate it.

fn sender(&self) -> ActorRef

Returns an Arc to the sender of the message being handled.

fn father(&self) -> ActorRef

Father of the actor.

fn children(&self) -> Vec<ActorRef>

Children of the actor.

fn monitoring(&self) -> Vec<ActorRef>

Lifecycle monitoring, list of monitored actors.

fn path(&self) -> Arc<ActorPath>

Logical path to the actor, such as /user/foo/bar/baz

fn identify_actor(&self, _name: String) -> Future<Option<ActorRef>, &'static str>

Tries to give an address from an actor path. Note that eventual futures are lazy, thus you need to await on the Future at some point, this makes this a synchronous call.