Struct robots::actors::props::Props

pub struct Props<Args: Arguments, A: Actor> { /* fields omitted */ }

Props is the current only ActorFactory.

It is used to generate actors in a reliable way, calling create will always create the same actor insatance, it is thus the way used to create actors through this crate.

Methods

impl<Args: Arguments, A: Actor> Props<Args, A>

fn new(
creator: Arc<Fn(Args) -> A + Sync + Send>,
args: Args
) -> Arc<ActorFactory>

Creates a Props which is a factory for A with the creator function and args args.